LCIA import with most features implemented
Instead of 2 different buttons, users now select the file type through a
radio button, and select their files through a single button
Leaving required fields empty now gives a warning dialog, instead of
just continuing
Name, description, and units for an impact category can now be filled in
by the user, and the correct function is then called.

However, after testing it with the test files given on the bw2 website:
https://2.docs.brightway.dev/intro.html#importing-lcia-methods-from-the-standard-excel-template
the import does not seem to do anything, while also not throwing a
warning or error. I'd have to have a look at what exactly is going wrong
here.